Re-work the output from the Status module.
Chanege the output to being an ordinary list & add styling for it
to the base style. I think that ideally I'd like this styling to be
part of the module, rather than the theme, but that functionality
doesn't exist yet.

Tweak the output formats to allow for richer responses, allowing for
some backwards compatibility.