Add a live preview of the chunk in the admin page.
Using the keydown event from the textarea we wait for either a pause
in the users typing, or after a max wait timeout. Once reached we fire
a call to the ajax funciton of the block for rendering of the current
text the user has entered & display the results in a prevously hidden
preview box.