Add the beginnings of a completely new Files module.
This files module is intended to repalce the current Docs module, which
is so full of spaghetti it's insane. I'm hoping that I can build it in
such a way that adding functionality by subclassing it will be easy.

I'm planning to build the new Gallery module in this manner. The
alternative might be to have this module support some form of dynamic
templating.

I've added a selection of icons from the Oxygen set
(http://www.oxygen-icons.org/), which are freely (and I believe suitably)
licensed, though this probably needs review.