Add the start of the new Gallery module.
It's based on the new Files module, so it's pretty simple since almost
all of the basic stuff is inherited. For the moment it doesn't actually
support resizing the images for preview, but the functionality that
overrides the Files module's linking to the images is present. Thus
at the moment it shows that this can work.

There are a number of things that this will need fixing in the main of
SquidgyCMS before release, notably proper dependencies, but for the
moment it works with the workarounds.