Rework how we figure out resizing support.
Use a combination of extensions (enabling easy filtering of pointless
candidates, txt files etc.) and mime types (for more accurate support
that won't be fooled by badly named files) to determine if a file can
be resized. This neccesitates changing the getSupportedTypes method to
pass back lists of each type, and as it's now being called a lot more
(every displayed file on the listings page) make it cache the result.

Since we're using the GalleryImage objects to get the MIME for the file
remove the error trigger from its constructor so that won't go recursive.
I'm not entirely happy about this, as it leaves that object without any
validation, and we end up making two of them for one image in the resize
callback.