Add an Inform module to improve action feedback to the user.
Eventually this will replace the print_success function, but as yet
I've not removed all the traces of that. For the moment it adds to
its use in the admin module, and replaces it as the default success
feedback responder.

The module use $_SESSION to store arrays of data which will be shown
to the user the next time that it is called to be shown. Note that
calling for a show returns the information as a string, rather than
printing it directly, and cleans the data. The data is cleaned so that
the session array doesn't grow indefinitely, and so the user doesn't
see the same messages over and over.