Remove most of the FSPHP stuff and the old Docs module that used it.
There's still some tidyup to be done, but this is a good start.
The functionality that this provided is now in the Files and Gallery
modules, and better integrated there.

Also remove most of the images that the Docs module used, keeping
only those for which there is as yet no counterpart in Files.